pyproject.toml
[project]
name = "kafka-trade-aggregator"
version = "0.1.0"
description = "Live trade feed aggregator using Kafka/Redpanda and Binance public WebSocket"
requires-python = ">=3.11"
dependencies = [
    "confluent-kafka>=2.3.0",
    "dotenv>=0.9.9",
    "faust-streaming>=0.10.0",
    "websocket-client>=1.7.0",
]

[dependency-groups]
dev = []
